+++ changelog.md
@@ -1,1 +1,7 @@
-Use this file to explain what has changed in your add-on since the previous release. This will be included automatically in the release description when used with GitHub actions.
+## Version 0.1
+
+Initial release.
+
+- Added detection of iCloud verification code popups
+- Identifies popup by window class (#32770) and iCloud text content
+- Extracts verification code from dialog (XXX XXX format)